Move exported components to root of visualizations
