+++ client/app/visualizations/Editor.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import PropTypes from "prop-types";
 import { EditorPropTypes } from "@/visualizations/prop-types";
-import registeredVisualizations from "@/visualizations";
+import registeredVisualizations from "@/visualizations/registeredVisualizations";
 
 export default function Editor({ type, ...otherProps }) {
   const { Editor } = registeredVisualizations[type];
