+++ client/app/visualizations/Renderer.jsx
@@ -3,7 +3,7 @@ import React, { useEffect, useRef } from "react";
 import PropTypes from "prop-types";
 import ErrorBoundary, { ErrorMessage } from "@/components/ErrorBoundary";
 import { RendererPropTypes } from "@/visualizations/prop-types";
-import registeredVisualizations from "@/visualizations";
+import registeredVisualizations from "@/visualizations/registeredVisualizations";
 
 export default function Renderer({
   type,
