--- client/app/visualizations/components/index.js
@@ -1,4 +1,0 @@
-import Renderer from "./Renderer";
-import Editor from "./Editor";
-
-export { Renderer, Editor };
