+++ client/app/visualizations/visualizationsSettings.js
